// src/socket/enum/socket-subscribe-message.enum.ts
export enum SubscribeMessageEnum {
  Toast = 'toast',
  ResetNotificationCount = 'reset-notification-count',
  Echo = 'echo',
  EchoGlobal = 'echo-global',
  Whoami = 'whoami',
  UserOut = 'user-out',
  SendPrivateData = 'private-data',
  SendPublishData = 'publish-data',
  ToastPrivate = 'toast-private',
  ToastPublish = 'toast-publish',
}